Fix compilation error

+++ doomclassic/doom/p_mobj.h
@@ -322,7 +322,7 @@ struct mobj_t
     
 };
 
-#define MF2_NEUTRAL_SPLASH int64(0x0000001000000000) // splash damage does not account for group
+#define MF2_NEUTRAL_SPLASH int(0x0000001000000000) // splash damage does not account for group
 
 #endif
